+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-8482 Added dedicated page for privacy policy content
 * VFS-7567 Added showing origin space in general shares sidebar
 * VFS-8038 Refactored build process to support faster builds on CI
 * VFS-8360 Fixed compatibility between Onezone GUI 21.02 and Oneprovider GUI 20.02 by moving shared properties caching to Oneprovider GUI